Move ::std::error to ::core::error.
- Moved entire module to libcore
- Apply `#[fundamental]` to the `Error` trait [breaking-change]
- Move Box downcast methods to methods on `Box<Error>` [breaking-change]
