+++ src/libstd/lib.rs
@@ -341,12 +341,13 @@ pub use core::result;
 #[stable(feature = "rust1", since = "1.0.0")]
 pub use core::option;
 
-pub mod error;
-
 #[stable(feature = "rust1", since = "1.0.0")]
 pub use alloc::boxed;
 #[stable(feature = "rust1", since = "1.0.0")]
 pub use alloc::rc;
+
+#[stable(feature = "rust1", since = "1.10.0")]
+pub use core::error;
 
 #[stable(feature = "rust1", since = "1.0.0")]
 pub use core_collections::borrow;
